requirements.txt
# Core
customtkinter>=5.2.2
Pillow>=10.0.0
requests>=2.31.0
urllib3>=2.0.0
beautifulsoup4>=4.12.0
python-owasp-zap-v2.4>=0.0.21

# Async / API
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0

# CLI
click>=8.1.7

# Config
PyYAML>=6.0.1

# Reporting (no external CDN needed — pure Python HTML)
# Optional: jinja2>=3.1.3  (only needed if you switch to template-based HTML)

# Dev / testing (not required at runtime)
# pytest>=8.0
# pytest-asyncio>=0.23
